refactor: extract f1 client into separate package
