+++ Cargo.toml
@@ -1,0 +1,38 @@
+[workspace]
+members = [
+    "crates/data",      # lib
+    "crates/client",    # lib
+    "crates/log",       # lib
+    "crates/live",      # bin
+    "crates/saver",     # bin
+    "crates/simulator", # bin
+]
+default-members = ["crates/live"]
+resolver = "2"
+
+exclude = ["api-backend"]
+
+[workspace.dependencies]
+data = { path = "crates/data" }
+client = { path = "crates/client" }
+log = { path = "crates/log" }
+
+tokio = { version = "1.37.0", features = ["full"] }
+tokio-stream = { version = "0.1.15", features = ["full"] }
+tracing = "0.1.40"
+tracing-subscriber = { version = "0.3.18", features = ["env-filter"] }
+
+axum = { version = "0.7.5", features = ["http2"] }
+tower-http = { version = "0.5.2", features = ["cors"] }
+tower_governor = "0.4.2"
+tokio-tungstenite = { version = "0.23.1", features = ["native-tls", "url"] }
+reqwest = { version = "0.12.4", features = ["native-tls"] }
+serde = { version = "1.0", features = ["derive"] }
+serde_json = { version = "1.0", features = ["raw_value"] }
+
+heck = "0.5.0"
+regex = "1.10.4"
+base64 = "0.22.1"
+flate2 = "1.0.30"
+futures = "0.3.30"
+dotenvy = "0.15.7"
