+++ crates/client/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "client"
+version = "0.3.0"
+edition = "2021"
+publish = false
+
+[lib]
+path = "src/client.rs"
+
+[dependencies]
+axum.workspace = true
+futures.workspace = true
+reqwest.workspace = true
+serde_json.workspace = true
+
+tokio.workspace = true
+tokio-stream.workspace = true
+tokio-tungstenite.workspace = true
+tracing.workspace = true
+
+data.workspace = true
